feature/model update v3 (#49)
# PR Context
We need to update our metadata model due to new insights deriving from
a) consulting more standards and schemas specific to the modelling of
health data, like upcoming profile "HealthDCAT-AP" in relation to the
development of the EHDS (European Health Data Space).
Further, we introduce a new entity called "bibliographic reference".
This new entity allows us to describe bibliographic references, e.g.
publications (books, book sections, journal articles). With that, we
added a new core entity to our metadata catalogue, which provides
valuable information about RKI's research.

# Added
- Entity _bibliographic reference_
- Vocabulary _bibliographic-reference-type_: a list of different
publication types.
- Vocabulary _resource-creation-method_: we want to have a better
distinction between the usage of _resourceTypeGeneral_ as a generic type
for resources and the context/method of how the resources were created.
- Vocabulary _personal-data_: we want to be able to say, if a resource
contains personal data or not.
- Added JSON Schema property _description_ to all entities and
properties, since we delete the specification of the model
(docs/index.md)

**resource**
- Property _conformsTo_: Providing information about standards used with
the resource.
- Property _doi_: For published resources.
- Property _hasLegalBasis_: Information about the legal basis on which
the data was processed; inspired by HealthDCAT-AP draft and [Data
Privacy Vocabulary](https://w3c.github.io/dpv/dpv/#hasLegalBasis)
- Property _hasPersonalData_: Information if the resource contains
personal data or not ;inspired by HealthDCAT-AP draft and [Data Privacy
Vocabulary](https://w3c.github.io/dpv/dpv/#hasPersonalData)
- Property _maxTypicalAge_ and _minTypicalAge_: Information about the
age of the researched population; inspired by HealthDCAT-AP draft
- Property _populationCoverage_: Information about the researched
population; inspired by HealthDCAT-AP draft
- Property _resourceCreationMethod_: Information about the context or
method used in order to create the resource

# Changes
all entities
- Property _identifierInPrimarySource_ has now a "minLength" and
"maxLength" and has a pattern to restrict newlines

**variable**
- Property _dataType_ does not expect a vocabulary item anymore, but a
string

**activity**
- Property _publication_ is now expecting the identifier of
_bibliographic-reference_ item

**resource**
- Property _publication_ is now expecting the identifier of a
*bibliographic reference* item
- Property _loincId_: new pattern

**distribution**
- Properties _accessURL_ and _downloadURL_ now have the type "array",
because there can be multiple repositories for accessing the
distribution (e.g. GitHub, Zenodo).
- Deleted some properties: We want only to have properties that are
relevant for the distribution's access. Hence, all properties concerning
actors involved in the creation of the resource were deleted.

**vocabularies**
- _theme_: It had too many concepts and was not suitable for our
purposes. We decided to delete most of the concepts and keep only the
main topics of RKI.
- _activity_: Deleted concepts due to discussions: We want to focus on
the distinction between third-party funded projects and internal
projects.
- _resource-type-general_: Deleted and added concepts due to
discussions: We have consulted other standards and schemas (e.g.
NFDI4Health, DataCite) to find more suitable resource types; although we
didn't find a perfect fit for the vocabulary, we found a few useful
concepts and added some of them.

**date properties** 
- Defined new patterns for dates: `YearMonthDay`, `YearMonth`, `Year`.
Delete `"format": "date"` and `"format": "datetime"` to ensure
validation.

**index.md**
Deleted all information that were relevant to the "specification" of the
model. We decided, that the maintenance of the specification is not
worth its benefit (mainly interoperability with standards). In the
future, it is more feasible and fitting to our use case to maintain
mappings to the portals we will contribute to.
But we want to keep the index file, because we may add some more
description on the schema on a more general level.

**README.md**
- Transferred contact and publisher information from _index.md_ to
README

# Removed
- Vocabulary _data-type.json_ because we decided against having a
controlled vocabulary for data types.

---------

Signed-off-by: FranziDiehr <26928727+fdiehr@users.noreply.github.com>
Signed-off-by: LeipoldL <60007921+LeipoldL@users.noreply.github.com>
Signed-off-by: erichesse <eric.hesse3@yahoo.de>
Co-authored-by: Nicolas Drebenstedt <897972+cutoffthetop@users.noreply.github.com>
Co-authored-by: LeipoldL <60007921+LeipoldL@users.noreply.github.com>
Co-authored-by: Kamran Ali <33874616+mr-kamran-ali@users.noreply.github.com>
Co-authored-by: erichesse <eric.hesse3@yahoo.de>
Co-authored-by: Nicolas Drebenstedt <drebenstedtn@rki.de>
Co-authored-by: rababerladuseladim <rababerladuseladim@users.noreply.github.com>